allow for using a temporary file as buffer
This is the main point of the fork, to use a temporary file instead of
an in-memory buffer for chunking the stream.

For most use cases, this isn't preferable, as it increases the amount of
I/O being done, along with CPU, in exchange for reducing the majority of
memory usage.

However, for use cases where the stream could be handling a lot of
memory moving through it, this could be preferable in order to reduce
the RAM usage while streaming.